+++ src/pattern/index.d.ts
@@ -1,3 +1,4 @@
 export {Throttle, IThrottleOptions} from './functions/throttle';
 export {IDisposable, Using} from './dispose/using';
 export {Create, CreateWithFactory} from './construct/construct';
+export {Retry, IRetryOptions} from './functions/retry';