+++ src/pattern/index.js
@@ -1,8 +1,9 @@
 /**
  * @module pattern
  */
+const {Retry} = require('./functions/retry');
 const {Using} = require('./dispose/using');
 const {Throttle} = require('./functions/throttle');
 const {Create, CreateWithFactory} = require('./construct/construct');
 
-module.exports = {Using, Throttle, CreateWithFactory, Create};
+module.exports = {Using, Throttle, CreateWithFactory, Create, Retry};